requirements.txt
Flask==0.12.1
paho-mqtt==1.3.0